+++ CodeRed/DirectX12/DirectX12FrameBuffer.cpp
@@ -12,26 +12,15 @@ using namespace CodeRed::DirectX12;
 
 CodeRed::DirectX12FrameBuffer::DirectX12FrameBuffer(
 	const std::shared_ptr<GpuLogicalDevice>& device,
-	const std::shared_ptr<GpuTexture>& render_target, 
-	const std::shared_ptr<GpuTexture>& depth_stencil) :
-	DirectX12FrameBuffer(device, 
-		render_target == nullptr ? nullptr : render_target->reference(), 
-		depth_stencil == nullptr ? nullptr : depth_stencil->reference())
-{
-	
-}
-
-CodeRed::DirectX12FrameBuffer::DirectX12FrameBuffer(
-	const std::shared_ptr<GpuLogicalDevice>& device,
-	const std::shared_ptr<GpuTextureRef>& render_target, 
+	const std::vector<std::shared_ptr<GpuTextureRef>>& render_targets,
 	const std::shared_ptr<GpuTextureRef>& depth_stencil) :
-	GpuFrameBuffer(device, render_target, depth_stencil)
+	GpuFrameBuffer(device, render_targets, depth_stencil)
 {
 	const auto dxDevice = static_cast<DirectX12LogicalDevice*>(mDevice.get())->device();
 
 	D3D12_DESCRIPTOR_HEAP_DESC rtvHeapInfo = {
 		D3D12_DESCRIPTOR_HEAP_TYPE_RTV,
-		1,
+		render_targets.empty() ? 1 : static_cast<UINT>(render_targets.size()),
 		D3D12_DESCRIPTOR_HEAP_FLAG_NONE,
 		0
 	};
@@ -53,69 +42,54 @@ CodeRed::DirectX12FrameBuffer::DirectX12FrameBuffer(
 		FailedException(DebugType::Create, { "ID3D12DescriptorHeap of Depth Stencil" })
 	);
 
-	DirectX12FrameBuffer::reset(render_target, depth_stencil);
-}
+	mRTVSize = dxDevice->GetDescriptorHandleIncrementSize(D3D12_DESCRIPTOR_HEAP_TYPE_RTV);
 
-void CodeRed::DirectX12FrameBuffer::reset(
-	const std::shared_ptr<GpuTextureRef>& render_target,
-	const std::shared_ptr<GpuTextureRef>& depth_stencil)
-{
-	//reset the render target and depth stencil
-	for (auto& rtv : mRenderTargets) rtv.reset();
-
-	mDepthStencil.reset();
-
-	const auto dxDevice = static_cast<DirectX12LogicalDevice*>(mDevice.get())->device();
-	
-	mRenderTargets[0] = render_target;
-	mDepthStencil = depth_stencil;
-	
 	//warning, when we create a frame buffer without rtv and dsv
 	//only output when we enable __ENABLE__CODE__RED__DEBUG__
 	CODE_RED_DEBUG_TRY_EXECUTE(
-		mRenderTargets[0] == nullptr &&
+		mRenderTargets.empty() &&
 		mDepthStencil == nullptr,
-		DebugReport::warning(DebugType::Create, 
+		DebugReport::warning(DebugType::Create,
 			{ "FrameBuffer" },
 			{ "there are no rtv and dsv." })
 	);
 
-	if (mRenderTargets[0] != nullptr) {
-
+	for (size_t index = 0; index < mRenderTargets.size(); index++) {
 		CODE_RED_DEBUG_THROW_IF(
-			mRenderTargets[0]->mipLevel().size() != 1 || mRenderTargets[0]->array().size() != 1,
+			mRenderTargets[index]->mipLevel().size() != 1 || mRenderTargets[index]->array().size() != 1,
 			"The size of mip level and array should be one, when we create a frame buffer."
 		);
-		
+
 		D3D12_RENDER_TARGET_VIEW_DESC view;
 
-		view.Format = enumConvert(mRenderTargets[0]->source()->format());
+		view.Format = enumConvert(mRenderTargets[index]->source()->format());
 
-		if (mRenderTargets[0]->source()->isArray()) {
+		if (mRenderTargets[index]->source()->isArray()) {
 
 			view.ViewDimension = D3D12_RTV_DIMENSION_TEXTURE2DARRAY;
-			view.Texture2DArray.FirstArraySlice = static_cast<UINT>(mRenderTargets[0]->array().Start);
-			view.Texture2DArray.MipSlice = static_cast<UINT>(mRenderTargets[0]->mipLevel().Start);
+			view.Texture2DArray.FirstArraySlice = static_cast<UINT>(mRenderTargets[index]->array().Start);
+			view.Texture2DArray.MipSlice = static_cast<UINT>(mRenderTargets[index]->mipLevel().Start);
 			view.Texture2DArray.PlaneSlice = 0;
 			view.Texture2DArray.ArraySize = 1;
-			
-		}else {
+
+		}
+		else {
 
 			view.ViewDimension = D3D12_RTV_DIMENSION_TEXTURE2D;
-			view.Texture2D.MipSlice = static_cast<UINT>(mRenderTargets[0]->mipLevel().Start);
+			view.Texture2D.MipSlice = static_cast<UINT>(mRenderTargets[index]->mipLevel().Start);
 			view.Texture2D.PlaneSlice = 0;
-			
+
 		}
 
 		dxDevice->CreateRenderTargetView(
-			std::static_pointer_cast<DirectX12Texture>(mRenderTargets[0]->source())->texture().Get(),
+			std::static_pointer_cast<DirectX12Texture>(mRenderTargets[index]->source())->texture().Get(),
 			&view,
-			mRenderTargetHeap->GetCPUDescriptorHandleForHeapStart()
+			{ mRenderTargetHeap->GetCPUDescriptorHandleForHeapStart().ptr + index * mRTVSize }
 		);
 	}
 
 	if (mDepthStencil != nullptr) {
-		
+
 		CODE_RED_DEBUG_THROW_IF(
 			mDepthStencil->mipLevel().size() != 1 || mDepthStencil->array().size() != 1,
 			"The size of mip level and array should be one, when we create a frame buffer."
@@ -133,11 +107,12 @@ void CodeRed::DirectX12FrameBuffer::reset(
 			view.Texture2DArray.MipSlice = static_cast<UINT>(mDepthStencil->mipLevel().Start);
 			view.Texture2DArray.ArraySize = 1;
 
-		}else {
+		}
+		else {
 
 			view.ViewDimension = D3D12_DSV_DIMENSION_TEXTURE2D;
 			view.Texture2D.MipSlice = static_cast<UINT>(mDepthStencil->mipLevel().Start);
-			
+
 		}
 
 		dxDevice->CreateDepthStencilView(