+++ CodeRed/DirectX12/DirectX12FrameBuffer.hpp
@@ -11,26 +11,21 @@ namespace CodeRed {
 	public:
 		explicit DirectX12FrameBuffer(
 			const std::shared_ptr<GpuLogicalDevice>& device,
-			const std::shared_ptr<GpuTexture>& render_target,
-			const std::shared_ptr<GpuTexture>& depth_stencil = nullptr);
-
-		explicit DirectX12FrameBuffer(
-			const std::shared_ptr<GpuLogicalDevice>& device,
-			const std::shared_ptr<GpuTextureRef>& render_target,
+			const std::vector<std::shared_ptr<GpuTextureRef>>& render_targets,
 			const std::shared_ptr<GpuTextureRef>& depth_stencil = nullptr);
-
+		
 		~DirectX12FrameBuffer() = default;
 
-		void reset(
-			const std::shared_ptr<GpuTextureRef>& render_target, 
-			const std::shared_ptr<GpuTextureRef>& depth_stencil) override;
-		
 		auto rtvHeap() const noexcept -> WRL::ComPtr<ID3D12DescriptorHeap> { return mRenderTargetHeap; }
 
 		auto dsvHeap() const noexcept -> WRL::ComPtr<ID3D12DescriptorHeap> { return mDepthStencilHeap; }
+
+		auto rtvSize() const noexcept -> size_t { return mRTVSize; }
 	private:
 		WRL::ComPtr<ID3D12DescriptorHeap> mRenderTargetHeap;
 		WRL::ComPtr<ID3D12DescriptorHeap> mDepthStencilHeap;
+
+		size_t mRTVSize = 0;
 	};
 	
 }