+++ CodeRed/DirectX12/DirectX12RenderPass.cpp
@@ -8,10 +8,11 @@ using namespace CodeRed::DirectX12;
 
 CodeRed::DirectX12RenderPass::DirectX12RenderPass(
 	const std::shared_ptr<GpuLogicalDevice>& device,
-	const std::optional<Attachment>& color, 
+	const std::vector<Attachment>& colors, 
 	const std::optional<Attachment>& depth) :
-	GpuRenderPass(device, color, depth)
+	GpuRenderPass(device, colors, depth)
 {
+	
 }
 
 #endif