+++ CodeRed/Vulkan/VulkanRenderPass.cpp
@@ -7,74 +7,73 @@ using namespace CodeRed::Vulkan;
 
 CodeRed::VulkanRenderPass::VulkanRenderPass(
 	const std::shared_ptr<GpuLogicalDevice>& device,
-	const std::optional<Attachment>& color,
+	const std::vector<Attachment>& colors, 
 	const std::optional<Attachment>& depth) :
-	GpuRenderPass(device, color, depth)
+	GpuRenderPass(device, colors, depth)
 {
-	const auto& colorAttachment = mColorAttachments[0];
-	const auto& depthAttachment = mDepthAttachment;
-	
 	size_t attachmentCount = 0;
-	
-	CODE_RED_TRY_EXECUTE(colorAttachment.has_value(), attachmentCount++);
-	CODE_RED_TRY_EXECUTE(depthAttachment.has_value(), attachmentCount++);
+
+	attachmentCount = attachmentCount + mColorAttachments.size();
+	attachmentCount = attachmentCount + (mDepthAttachment.has_value() ? 1 : 0);
 	
 	std::vector<vk::AttachmentDescription> attachments(attachmentCount);
-
+	std::vector<vk::AttachmentReference> colorsReference(mColorAttachments.size());
+	vk::AttachmentReference depthReference;
+	
 	attachmentCount = 0;
-	
+
+	for (size_t index = 0; index < mColorAttachments.size(); index++) {
+		const auto& colorAttachment = mColorAttachments[index];
+
+		colorsReference[index].attachment = static_cast<uint32_t>(attachmentCount);
+		colorsReference[index].layout = vk::ImageLayout::eColorAttachmentOptimal;
+		
+		attachments[attachmentCount++]
+			.setFlags(vk::AttachmentDescriptionFlags(0))
+			.setFormat(enumConvert(colorAttachment.Format))
+			.setSamples(vk::SampleCountFlagBits::e1)
+			.setLoadOp(enumConvert(colorAttachment.Load))
+			.setStoreOp(enumConvert(colorAttachment.Store))
+			.setStencilLoadOp(enumConvert(colorAttachment.StencilLoad))
+			.setStencilStoreOp(enumConvert(colorAttachment.StencilStore))
+			.setInitialLayout(enumConvert(colorAttachment.InitialLayout))
+			.setFinalLayout(enumConvert(colorAttachment.FinalLayout));
+	}
+
 	CODE_RED_TRY_EXECUTE(
-		colorAttachment.has_value(),
+		mDepthAttachment.has_value(),
 		attachments[attachmentCount++]
 		.setFlags(vk::AttachmentDescriptionFlags(0))
-		.setFormat(enumConvert(colorAttachment->Format))
+		.setFormat(enumConvert(mDepthAttachment->Format))
 		.setSamples(vk::SampleCountFlagBits::e1)
-		.setLoadOp(enumConvert(colorAttachment->Load))
-		.setStoreOp(enumConvert(colorAttachment->Store))
-		.setStencilLoadOp(enumConvert(colorAttachment->StencilLoad))
-		.setStencilStoreOp(enumConvert(colorAttachment->StencilStore))
-		.setInitialLayout(enumConvert(colorAttachment->InitialLayout))
-		.setFinalLayout(enumConvert(colorAttachment->FinalLayout))
+		.setLoadOp(enumConvert(mDepthAttachment->Load))
+		.setStoreOp(enumConvert(mDepthAttachment->Store))
+		.setStencilLoadOp(enumConvert(mDepthAttachment->StencilLoad))
+		.setStencilStoreOp(enumConvert(mDepthAttachment->StencilStore))
+		.setInitialLayout(enumConvert(mDepthAttachment->InitialLayout))
+		.setFinalLayout(enumConvert(mDepthAttachment->FinalLayout))
 	);
 
 	CODE_RED_TRY_EXECUTE(
-		depthAttachment.has_value(),
-		attachments[attachmentCount++]
-		.setFlags(vk::AttachmentDescriptionFlags(0))
-		.setFormat(enumConvert(depthAttachment->Format))
-		.setSamples(vk::SampleCountFlagBits::e1)
-		.setLoadOp(enumConvert(depthAttachment->Load))
-		.setStoreOp(enumConvert(depthAttachment->Store))
-		.setStencilLoadOp(enumConvert(depthAttachment->StencilLoad))
-		.setStencilStoreOp(enumConvert(depthAttachment->StencilStore))
-		.setInitialLayout(enumConvert(depthAttachment->InitialLayout))
-		.setFinalLayout(enumConvert(depthAttachment->FinalLayout))
+		mDepthAttachment.has_value(),
+		depthReference
+		.setAttachment(static_cast<uint32_t>(colorsReference.size()))
+		.setLayout(vk::ImageLayout::eDepthStencilAttachmentOptimal)
 	);
 	
-	vk::AttachmentReference colorReference = {};
-	vk::AttachmentReference depthReference = {};
-
-	colorReference.attachment = 0;
-	colorReference.layout = vk::ImageLayout::eColorAttachmentOptimal;
-
-	//if we do not have the color attachment the reference of depth is 0
-	//we only support one render target in this version
-	depthReference.attachment = color.has_value() ? 1 : 0;
-	depthReference.layout = vk::ImageLayout::eDepthStencilAttachmentOptimal;
-
 	vk::SubpassDescription subPassInfo = {};
 
 	subPassInfo
 		.setFlags(vk::SubpassDescriptionFlags(0))
 		.setPipelineBindPoint(vk::PipelineBindPoint::eGraphics)
 		.setInputAttachmentCount(0)
-		.setColorAttachmentCount(colorAttachment.has_value() ? 1 : 0)
+		.setColorAttachmentCount(static_cast<uint32_t>(colorsReference.size()))
 		.setPreserveAttachmentCount(0)
 		.setPInputAttachments(nullptr)
-		.setPColorAttachments(colorAttachment.has_value() ? &colorReference : nullptr)
+		.setPColorAttachments(colorsReference.data())
 		.setPPreserveAttachments(nullptr)
 		.setPResolveAttachments(nullptr)
-		.setPDepthStencilAttachment(depthAttachment.has_value() ? &depthReference : nullptr);
+		.setPDepthStencilAttachment(mDepthAttachment.has_value() ? &depthReference : nullptr);
 
 	vk::RenderPassCreateInfo info = {};
 
@@ -82,14 +81,14 @@ CodeRed::VulkanRenderPass::VulkanRenderPass(
 		.setPNext(nullptr)
 		.setFlags(vk::RenderPassCreateFlags(0))
 		.setAttachmentCount(static_cast<uint32_t>(attachments.size()))
-		.setPAttachments(attachments.empty() ? nullptr : attachments.data())
+		.setPAttachments(attachments.data())
 		.setSubpassCount(1)
 		.setPSubpasses(&subPassInfo)
 		.setDependencyCount(0)
 		.setPDependencies(nullptr);
 
 	const auto vkDevice = std::static_pointer_cast<VulkanLogicalDevice>(mDevice)->device();
-	
+
 	mRenderPass = vkDevice.createRenderPass(info);
 }